requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
billiard==3.6.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
dill==0.3.4
Django==4.0.3
docutils==0.18.1
frozenlist==1.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kombu==5.2.4
multidict==6.0.2
multiprocess==0.70.12.2
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pypiwin32==223
pyTelegramBotAPI==4.4.0
python-decouple==3.6
pytz==2021.3
pywin32==303
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wincertstore==0.2
yarl==1.7.2
zope.interface==5.4.0
